More Progress on #45, and progress towards issue #48
A Couple of new features have been added. The User profile is now shown on the profile page, along with Username and Email. No option to customize or choose it yet, so a default profile is automatically made when a user registers and gives it a profile of default.png.