Classes Page
Added a Classes page navigatable from the nav bar at the top. Upon visiting the page, it asks to enter a class code. One you enter the code (Only one at this point is ABCDEF), a valid one will add you to the classroom and an invalid one will result in an error.